% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_simple_attr}
\alias{add_simple_attr}
\title{Add attributes to dataset from single-value variables in another dataset.}
\usage{
add_simple_attr(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Add attributes to dataset from single-value variables in another dataset.
}
