% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_aux_values.R
\name{adjust_aux_values}
\alias{adjust_aux_values}
\title{Adjust auxiliary data}
\usage{
adjust_aux_values(svy_year, aux_df, value_var)
}
\arguments{
\item{svy_year}{numeric: A vector with the decimal survey year.}

\item{aux_df}{data.frame: A data frame containing the variable to be retrieved.}

\item{value_var}{character: The variable to be adjusted.}
}
\value{
numeric
}
\description{
Adjust auxiliary statistics for surveys that span multiple calender years.
}
\details{
GDP, PCE and other auxiliary statistics are available for calendar years.
Some surveys do however span over two calendar years. This function can be
used to adjust for the corresponding mismatch between calendar and survey
year.

Values are adjusted by the weighted average of the years in question.
}
\keyword{internal}
