% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_welfare_mean.R
\name{adjust_welfare_mean}
\alias{adjust_welfare_mean}
\alias{safe_adjust_welfare_mean}
\title{Adjust welfare mean}
\usage{
adjust_welfare_mean(svy_table, ref_pce, ref_gdp, method)

safe_adjust_welfare_mean(...)
}
\arguments{
\item{svy_table}{data.frame: A table with surveys to use for the line-up.
Output of \code{\link[=get_closest_surveys]{get_closest_surveys()}}.}

\item{ref_pce}{numeric: PCE value for the reference year.}

\item{ref_gdp}{numeric: GDP value for the reference year.}

\item{method}{character: A string with the adjustment method to use. Output
of \code{\link[=get_welfare_adjustment_method]{get_welfare_adjustment_method()}}.}
}
\value{
\code{data.table}
}
\description{
Adjust the welfare mean according to the correct interpolation or
extrapolation method.
}
\details{
Interpolation and extrapolation of the welfare mean is computed differently
depending on various parameters. The adjustment depends on whether surveys
are available on one side or both sides of the requested year, the
availability of GDP and PCE data as well as the World Bank region the
specified country belongs to. This function computes the adjusted mean by
applying the specified method.

See
\href{http://documents1.worldbank.org/curated/en/664751553100573765/pdf/135460-WP-PUBLIC-Disclosed-3-21-2019.pdf}{Prydz et. al (2019)}
for more detailed documentation.
}
\references{
Prydz, E.B, Jolliffe, D., Lakner, C., Mahler, D.G., Sangraula, P. (2019)
'National Accounts Data used in Global Poverty Measurement'.
\emph{Global Poverty Monitoring Technical Note}, 8.
}
\keyword{internal}
