% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lcu_table.R
\name{calculate_survey_mean}
\alias{calculate_survey_mean}
\title{Calculate survey mean}
\usage{
calculate_survey_mean(inv)
}
\arguments{
\item{inv}{character: Survey ID.}
}
\value{
data.table
}
\description{
Calculate survey mean in local currency units (LCU) for a single survey.
}
\keyword{internal}
