% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_dist_stats.R
\name{compute_dist_stats}
\alias{compute_dist_stats}
\title{compute_dist_stats}
\usage{
compute_dist_stats(dt, mean)
}
\arguments{
\item{dt}{data.table: A survey dataset.}

\item{mean}{numeric: A value with the survey mean.}
}
\value{
list
}
\description{
compute_dist_stats
}
\keyword{internal}
