% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_mld}
\alias{compute_mld}
\title{Mean log deviation}
\usage{
compute_mld(measure, weight)
}
\arguments{
\item{measure}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}
}
\value{
numeric
}
\description{
Compute mean log deviation.
}
\details{
Given a vector of income or consumption values and their respective weights
\code{compute_mld()} computes the mean log deviation for the distribution.

Note that the input parameters cannot contain missing values and that
\code{measure} must be sorted in increasing order.
}
\examples{
# Simple example
compute_mld(measure = 1:100, weight = rep(1, 100))

# Microdata example
data("microdata")
compute_mld(measure = microdata$welfare, microdata$weight)

}
\seealso{
\code{\link[=compute_stats]{compute_stats()}} \code{\link[=compute_gini]{compute_gini()}} \code{\link[=compute_lorenz]{compute_lorenz()}}
}
