% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{Summary statistics}
\usage{
compute_stats(y, w)
}
\arguments{
\item{y}{numeric: A vector of income or consumption values}

\item{w}{numeric: A vector of weights}
}
\value{
Returns a list of summary statistics.
}
\description{
Compute summary statistics, including Gini, Lorenz curve, weighted mean and
minimum and maximum values.
}
\details{
Note that the input parameters cannot contain missing values and that
\code{y} must be sorted in increasing order.
}
