% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cache_file.R
\name{create_cache_file}
\alias{create_cache_file}
\title{Cache survey data}
\usage{
create_cache_file(
  pipeline_inventory = NULL,
  pip_data_dir = gls$PIP_DATA_DIR,
  tool = c("PC", "TB"),
  cache_svy_dir = gls$CACHE_SVY_DIR_PC,
  compress = 100,
  verbose = getOption("pipdm.verbose"),
  force = FALSE,
  cpi_dt = pipload::pip_load_aux("cpi"),
  ppp_dt = pipload::pip_load_aux("ppp"),
  pfw_dt = pipload::pip_load_aux("pfw"),
  pop_dt = pipload::pip_load_aux("pop")
)
}
\arguments{
\item{pipeline_inventory}{data.table: Pipeline inventory table.}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{tool}{character: Either "PC" or "TB" for poverty calculator and Table
Baker, receptively}

\item{cache_svy_dir}{character: Output folder for the cached survey data.}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{verbose}{boolean: If TRUE additional messages are printed to the
console.}

\item{force}{logical: if TRUE, all files will be generate again.}

\item{cpi_dt}{data frame with CPI data to deflate welfare}

\item{ppp_dt}{data frame with PPP data to deflate welfare}

\item{pfw_dt}{data frame with Price FrameWork data}
}
\value{
list: creation status and cache data availability
}
\description{
Save a cleaned version of the survey data in the cache survey data directory.
}
