% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_adjust_nac_values}
\alias{db_adjust_nac_values}
\title{Adjust national account values}
\usage{
db_adjust_nac_values(df)
}
\arguments{
\item{df}{data.table: Output of \code{\link[=db_merge_anchor_nac]{db_merge_anchor_nac()}}.}
}
\value{
\code{data.table}
}
\description{
Adjust GDP and PCE values for surveys that span multiple calender
years.
}
\details{
Values are adjusted by the weighted average of the years in question.
}
\seealso{
\code{\link[=adjust_aux_values]{adjust_aux_values()}}
}
\keyword{internal}
