% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_adjust_welfare_mean}
\alias{db_adjust_welfare_mean}
\title{Adjust the welfare mean for all reference years}
\usage{
db_adjust_welfare_mean(df)
}
\arguments{
\item{df}{data.table: Output of \code{\link[=db_get_closest_surveys]{db_get_closest_surveys()}}.}
}
\value{
\code{data.table}
}
\description{
Adjust the welfare mean for all countries and reference years.
}
\seealso{
\code{\link[=get_welfare_adjustment_method]{get_welfare_adjustment_method()}} \code{\link[=adjust_welfare_mean]{adjust_welfare_mean()}}
}
\keyword{internal}
