% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clean_aux.R
\name{db_clean_aux}
\alias{db_clean_aux}
\title{Clean auxiliary data}
\usage{
db_clean_aux(dt, name, pip_years)
}
\arguments{
\item{dt}{data.table: An auxiliary dataset.}

\item{name}{character: Name, e.g. 'ppp', 'gdp' etc.}

\item{pip_years}{numeric: A vector with calender years used in PIP.}
}
\value{
data.table
}
\description{
Clean auxiliary data for PIP webpage.
}
