% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_dist_stats.R
\name{db_compute_dist_stats}
\alias{db_compute_dist_stats}
\title{Calculate distributional statistics}
\usage{
db_compute_dist_stats(dt, mean, pop_table, cache_id, gc = FALSE)
}
\arguments{
\item{dt}{data.table: A survey dataset.}

\item{mean}{numeric: A value with the survey mean.}

\item{pop_table}{data.table: A table with population data.}

\item{cache_id}{character: cache id to identify the right process}

\item{gc}{logical: If TRUE garbage collection is forced.}
}
\value{
list
}
\description{
Calculate distributional statistics  for a single survey dataset.
}
