% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_lorenz.R
\name{db_compute_lorenz}
\alias{db_compute_lorenz}
\title{Compute Lorenz}
\usage{
db_compute_lorenz(dt)
}
\arguments{
\item{dt}{data.table: A survey dataset.}
}
\description{
Calculate points on the Lorenz curve for a single survey dataset.
}
\details{
Only available for microdata.
}
