% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_predicted_means.R
\name{db_compute_predicted_means}
\alias{db_compute_predicted_means}
\title{Calculate predicted means}
\usage{
db_compute_predicted_means(dt)
}
\arguments{
\item{dt}{data.table: Output of \code{\link[=db_get_closest_surveys]{db_get_closest_surveys()}}.}
}
\value{
\code{data.table}
}
\description{
Predict welfare means for all reference years.
}
\keyword{internal}
