% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_censoring_table.R
\name{db_create_censoring_table}
\alias{db_create_censoring_table}
\title{Create censoring table}
\usage{
db_create_censoring_table(censored, coverage_table, coverage_threshold)
}
\arguments{
\item{censored}{list: List with manual provided countries or regions to censor.}

\item{coverage_table}{data.table: Coverage table. Output of \code{db_create_coverage_table()}.}

\item{coverage_threshold}{numeric: Value with the threshold to censor regions based on their coverage (e.g 50 \%).}
}
\description{
Create censoring table based on manual inputs recorded in the AUX data, as well as the regional coverage.
}
