% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_companion_list.R
\name{db_create_companion_list}
\alias{db_create_companion_list}
\title{Create companion list of all information relavan for each survey data}
\usage{
db_create_companion_list(
  df,
  cpi_table = pipload::pip_load_aux("cpi"),
  ppp_table = pipload::pip_load_aux("ppp"),
  pfw_table = pipload::pip_load_aux("pfw"),
  pop_table = pipload::pip_load_aux("pop")
)
}
\arguments{
\item{pop_table}{}
}
\value{

}
\description{
Create companion list of all information relavan for each survey data
}
