% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_coverage_table.R
\name{db_create_coverage_table}
\alias{db_create_coverage_table}
\title{Create coverage table}
\usage{
db_create_coverage_table(
  ref_year_table,
  pop_table,
  ref_years,
  digits = 2,
  special_countries = c("ARG", "CHN", "IDN", "IND")
)
}
\arguments{
\item{ref_year_table}{data.table: Full interpolated means table. Output of
\code{db_create_ref_year_table()}.}

\item{pop_table}{data.table: A table with population data.}

\item{ref_years}{numeric: A vector with reference years.}

\item{digits}{numeric: The number of digits the returned coverage numbers are
rounded by.}

\item{special_countries}{character: A string with 3-letter country codes.}
}
\description{
Create a table with coverage estimates at regional, WLD and TOT levels.
}
