% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_decomposition_table.R
\name{db_create_decomposition_table}
\alias{db_create_decomposition_table}
\title{Create decomposition table}
\usage{
db_create_decomposition_table(dsm_table)
}
\arguments{
\item{dsm_table}{data.table: A table with deflated survey means.}
}
\value{
data.table
}
\description{
Create decomposition table
}
