% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dist_table.R
\name{db_create_dist_table}
\alias{db_create_dist_table}
\title{Create distributional statistics table}
\usage{
db_create_dist_table(dl)
}
\arguments{
\item{dl}{list: A list with distributional statistics datasets.}
}
\value{
data.table
}
\description{
Create distributional statistics table
}
