% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dsm_table.R
\name{db_create_dsm_table}
\alias{db_create_dsm_table}
\title{Create deflated survey mean table}
\usage{
db_create_dsm_table(
  inv = NULL,
  cpi = NULL,
  ppp = NULL,
  dt = NULL,
  append = TRUE,
  datadir = getOption("pip.datadir"),
  pipedir = getOption("pip.pipedir")
)
}
\arguments{
\item{inv}{character: Survey IDs to be parsed to \code{db_create_table()} in case
argument \code{dt} is NULL. Only works if \code{dt} is NULL.}

\item{cpi}{data.table: A table with CPI data.}

\item{ppp}{data.table: A table with PPP data.}

\item{dt}{data.table: An updated table with LCU survey means. If NULL calls
\code{db_create_lcu_table()} parsing argument \code{inv}.}

\item{append}{logical: If TRUE, append to current DSM table file and save.
TRUE is default.}

\item{datadir}{character: Path to the PIP-Data directory. Defaults to
\code{getOption("pip.datadir")}.}

\item{pipedir}{character: Path to the ingestion pipeline directory. Defaults
to \code{getOption("pip.pipedir")}.}
}
\value{
data.table
}
\description{
Create a table with deflated welfare means for each country and surveyid
year.
}
