% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_estimation_table.R
\name{db_create_estimation_table}
\alias{db_create_estimation_table}
\title{Finalize reference year table}
\usage{
db_create_estimation_table(ref_year_table, dist_table)
}
\arguments{
\item{ref_year_table}{data.table: Output of \code{\link[=db_create_ref_year_table]{db_create_ref_year_table()}}.}

\item{dist_table}{data.table: Output of \code{\link[=db_create_dist_table]{db_create_dist_table()}}.}
}
\value{
data.table
}
\description{
Finalize the reference year table by subsetting the correct rows and
columns.
}
