% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_gd_svy_mean_table.R
\name{db_create_gd_svy_mean_table}
\alias{db_create_gd_svy_mean_table}
\title{Create table with grouped data survey means}
\usage{
db_create_gd_svy_mean_table(pcn_master_path, pfw_table, inventory)
}
\arguments{
\item{pcn_master_path}{character: Path to PCN masterfile.}

\item{pfw_table}{data.table: A table with the price framework file.}

\item{inventory}{data.table: A table with the PIP inventory.}
}
\description{
Create table with grouped data survey means from the PCN Masterfile.
}
\details{
Survey means cannot be automatically calculated for grouped data, so at some
stage the mean needs to be entered manually. This function reads from the PCN
masterfile to ensure that PCN and PIP uses the same data means.

This source \strong{should} be changed in the future.
}
