% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lcu_table.R
\name{db_create_lcu_table}
\alias{db_create_lcu_table}
\title{Create survey mean table}
\usage{
db_create_lcu_table(inv = NULL)
}
\arguments{
\item{inv}{character: vector of survey IDs. If NULL the entire repository
will be used.}
}
\value{
data.table
}
\description{
Create a table with welfare means in Local Currency Units (LCU) for each survey.
}
