% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lkup_table.R
\name{db_create_lkup_table}
\alias{db_create_lkup_table}
\title{Create reference year look-up table}
\usage{
db_create_lkup_table(dt, nac_table, ref_years)
}
\arguments{
\item{dt}{data.table: A table with survey data. See details.}

\item{nac_table}{data.table: A table with GDP and PCE data. Output of
\code{\link[=db_subset_nac_table]{db_subset_nac_table()}}.}

\item{ref_years}{numeric: Vector with reference years.}
}
\value{
\code{data.table}
}
\description{
Create a look-up table with information on all reference years and surveys
for further data manipulation.
}
\details{
\code{dt} must included information on survey metadata, survey national account
values and deflated survey means.
}
\keyword{internal}
