% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_metadata_table.R
\name{db_create_metadata_table}
\alias{db_create_metadata_table}
\title{Create metadata table}
\usage{
db_create_metadata_table(pfw_table, lcu_table)
}
\arguments{
\item{pfw_table}{data.table: A table with the price framework file.}

\item{lcu_table}{data.table: A table with newly estimated LCU survey means.
Output of \code{db_create_lcu_table()}.}
}
\description{
Create metadata table
}
