% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_estimation_table.R
\name{db_create_ref_estimation_table}
\alias{db_create_ref_estimation_table}
\title{Create reference estimation table}
\usage{
db_create_ref_estimation_table(ref_year_table, dist_table)
}
\arguments{
\item{ref_year_table}{data.table: Output of \code{\link[=db_create_ref_year_table]{db_create_ref_year_table()}}.}

\item{dist_table}{data.table: Output of \code{\link[=db_create_dist_table]{db_create_dist_table()}}.}
}
\value{
data.table
}
\description{
Merge the interpolated means and distributional statistics tables.
}
