% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_create_ref_year_table}
\alias{db_create_ref_year_table}
\title{Create reference year table}
\usage{
db_create_ref_year_table(
  gdp_table,
  pce_table,
  svy_anchor,
  svy_mean_table = NULL,
  ref_years = pipdm:::reference_years,
  pip_years = pipdm:::pip_years
)
}
\arguments{
\item{gdp_table}{data.frame: A table with GDP data.}

\item{pce_table}{data.frame: A table with PCE data.}

\item{svy_anchor}{data.frame: A table with survey metadata information.}

\item{svy_mean_table}{data.frame: A table with survey means.}

\item{ref_years}{numeric: a vector with reference years.}

\item{pip_years}{numeric: A vector with calender years used in PIP.}
}
\value{
\code{data.frame}
}
\description{
Create a table with adjusted welfare means for each reference year.
}
\details{
\code{db_create_ref_year_table()} creates a table for with interpolated or
extrapolated welfare means for all specified reference years based on the
provided survey, GDP and PCE data.
}
\seealso{
\code{\link[=adjust_welfare_mean]{adjust_welfare_mean()}}
}
