% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_svy_anchor.R
\name{db_create_svy_anchor}
\alias{db_create_svy_anchor}
\title{Create survey anchor}
\usage{
db_create_svy_anchor(df)
}
\arguments{
\item{df}{data.frame: The price framework file.}
}
\value{
\code{data.table}
}
\description{
Create survey anchor
}
\keyword{internal}
