% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_svy_estimation_table.R
\name{db_create_svy_estimation_table}
\alias{db_create_svy_estimation_table}
\title{Create survey estimation table}
\usage{
db_create_svy_estimation_table(dsm_table, dist_table)
}
\arguments{
\item{dsm_table}{data.table: A table with deflated survey means.}

\item{dist_table}{data.table: Output of \code{\link[=db_create_dist_table]{db_create_dist_table()}}.}
}
\value{
data.table
}
\description{
Merge the survey mean and distributional statistics tables.
}
