% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_filter_inventory.R
\name{db_filter_inventory}
\alias{db_filter_inventory}
\title{filter repository of those surveys that are already in used in dsm_table}
\usage{
db_filter_inventory(
  raw_inventory = NULL,
  datadir = getOption("pip.datadir"),
  pipedir = getOption("pip.pipedir")
)
}
\arguments{
\item{raw_inventory}{dataframe from \code{paste0(datadir, "_inventory/inventory.fst")}}

\item{datadir}{character: PIP-Data directory}

\item{pipedir}{character: Pipeline ingestion directory}
}
\value{

}
\description{
filter repository of those surveys that are already in used in dsm_table
}
