% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_finalize_ref_year_table}
\alias{db_finalize_ref_year_table}
\title{Finalize reference year table}
\usage{
db_finalize_ref_year_table(df, svy_anchor)
}
\arguments{
\item{df}{data.table: Output of \code{\link[=db_select_lineup_surveys]{db_select_lineup_surveys()}}.}

\item{svy_anchor}{data.frame: A table with survey metadata information.}
}
\value{
\code{data.table}
}
\description{
Finalize the reference year table by subsetting the correct rows and
columns.
}
\keyword{internal}
