% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_closest_surveys.R
\name{db_get_closest_surveys}
\alias{db_get_closest_surveys}
\title{Get closest surveys for all reference years}
\usage{
db_get_closest_surveys(dt)
}
\arguments{
\item{dt}{data.table: Output of \code{\link[=db_create_lkup_table]{db_create_lkup_table()}}.}
}
\value{
\code{data.table}
}
\description{
Retrieve the closest surveys for each reference year.
}
\seealso{
\code{\link[=get_closest_surveys]{get_closest_surveys()}}
}
\keyword{internal}
