% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_and_clean.R
\name{db_load_and_clean}
\alias{db_load_and_clean}
\title{Load and clean welfare data from vector of survey ID}
\usage{
db_load_and_clean(survey_id, maindir)
}
\arguments{
\item{survey_id}{character: vector of survey ID from
\code{pipload::pip_find_data(filter_to_pc = TRUE)}}

\item{maindir}{character: Main directory path}
}
\value{

}
\description{
Load and clean welfare data from vector of survey ID
}
