% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_merge_anchor_nac.R
\name{db_merge_anchor_nac}
\alias{db_merge_anchor_nac}
\title{Merge survey anchor with national accounts data}
\usage{
db_merge_anchor_nac(svy_anchor, nac_table)
}
\arguments{
\item{svy_anchor}{data.table: A table with survey metadata information.}

\item{nac_table}{data.table: Output of \code{\link[=db_subset_nac_table]{db_subset_nac_table()}}.}
}
\value{
\code{data.table}
}
\description{
Create a combined table of survey metadata information and national accounts
data.
}
\details{
For surveys that span two calender years the returned GDP and PCE values are
adjusted by the weighted average of the years in question.
}
\seealso{
\code{\link[=adjust_aux_values]{adjust_aux_values()}}
}
\keyword{internal}
