% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ppp_list.R
\name{db_ppp_list}
\alias{db_ppp_list}
\title{Create generic list of PPP version according the ppp year, release version
and adaptation version}
\usage{
db_ppp_list(ppp)
}
\arguments{
\item{ppp}{data frame with ppp value. must have the same strcuture as in
\code{pipload::pip_load_aux("ppp")}}
}
\value{
list of ppp data frames
}
\description{
Create generic list of PPP version according the ppp year, release version
and adaptation version
}
