% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_select_lineup_surveys.R
\name{db_select_lineup_surveys}
\alias{db_select_lineup_surveys}
\title{Select correct line-up surveys}
\usage{
db_select_lineup_surveys(dt)
}
\arguments{
\item{dt}{data.table: Output of \code{\link[=db_get_closest_surveys]{db_get_closest_surveys()}}.}
}
\value{
data.table
}
\description{
Select a single line-up survey when there are multiple line-up surveys
available.
}
\seealso{
\code{\link[=select_lineup_survey]{select_lineup_survey()}}
}
\keyword{internal}
