% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_nac_table.R
\name{db_subset_nac_table}
\alias{db_subset_nac_table}
\title{Subset national accounts data}
\usage{
db_subset_nac_table(nac_table, pip_years)
}
\arguments{
\item{nac_table}{data.table: Output of \code{\link[=db_create_nac_table]{db_create_nac_table()}}}

\item{pip_years}{numeric: A vector with calender years used in PIP.}
}
\value{
\code{data.table}
}
\description{
Subset the output of \code{db_create_nac_table()} to only include years used by
the Poverty and Inequality Platform (PIP).
}
\keyword{internal}
