% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_old_file.R
\name{delete_old_file}
\alias{delete_old_file}
\title{Delete cache files that are no longer in use}
\usage{
delete_old_file(
  target_dir = "in",
  gls,
  cache_ids = NULL,
  verbose = FALSE,
  delete = NULL
)
}
\arguments{
\item{target_dir}{character: directory where the cache files are stored. If
could be a patch, like \code{gls$OUT_SVY_DIR_PC} from}

\item{gls}{list: List of globals.}

\item{cache_ids}{character:vector with  Cache_ids. default from
\code{db_create_pipeline_inventory()}
\code{pipload::add_gls_to_env()} or it could be one of the two shortcuts, "in"
or "out". The former refers to \code{gls$CACHE_SVY_DIR_PC}. the latter refers to
\code{gls$OUT_SVY_DIR_PC}.}

\item{verbose}{logical: whether to display messages}

\item{delete}{logical or NULL. If Null, menu with choices will be displayed
(default). If false, no file will be deleleted. If TRUE, all old files not
available in cache_id will be deleted}
}
\value{
invisible NA
}
\description{
Delete cache files that are no longer in use
}
