% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distance_weight}
\alias{distance_weight}
\title{get weights from survey year to reference year}
\usage{
distance_weight(ref_year, svy_years)
}
\arguments{
\item{ref_year}{numeric: reference year}

\item{svy_years}{numeric: survey years. Length either 1 or 2}
}
\value{
numeric vector with both weights that sum up to 1
}
\description{
get weights from survey year to reference year
}
