% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter2list}
\alias{filter2list}
\title{FIlter data.frame and convert it to list}
\usage{
filter2list(dt_, ...)
}
\arguments{
\item{dt_}{data.frame}

\item{condition}{condition in \code{i}  as in  \code{DT[i,j]}}
}
\value{
list
}
\description{
FIlter data.frame and convert it to list
}
\examples{
library(data.table)
dt <- data.table(x = c("a", "a", "b", "c", "c"),
 y = 1:5)

filter2list(dt, y > 3)
}
