% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_new_svy_data.R
\name{find_new_svy_data}
\alias{find_new_svy_data}
\title{Find which data is not available in cache folder or has changed SVY ID}
\usage{
find_new_svy_data(cache_id, filename, tool = c("PC", "TB"), cache_svy_dir)
}
\arguments{
\item{cache_id}{character: vector with cache Ids from pipeline inventory}

\item{filename}{character: vector with original dta file names}

\item{tool}{character: Either "PC" or "TB" for poverty calculator and Table
Baker, receptively}

\item{cache_svy_dir}{character: Output folder for the cached survey data.}
}
\description{
Find which data is not available in cache folder or has changed SVY ID
}
