% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{get_closest_surveys}
\alias{get_closest_surveys}
\title{Get closest surveys to a reference year}
\usage{
get_closest_surveys(svy_lkup, ref_year)
}
\arguments{
\item{svy_lkup}{data.frame: Look-up table for survey data.}

\item{ref_year}{numeric: The selected reference year.}
}
\value{
\code{data.frame}
}
\description{
For any reference year, retrieves information about the closest surveys.
}
\keyword{internal}
