% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup-utils.R
\name{get_welfare_adjustment_method}
\alias{get_welfare_adjustment_method}
\title{Get welfare adjustment method}
\usage{
get_welfare_adjustment_method(svy_table, ref_year, ref_gdp, ref_pce)
}
\arguments{
\item{svy_table}{data.frame: A table with information on surveys and
national accounts data.}

\item{ref_year}{numeric: A value with the reference year.}

\item{ref_gdp}{numeric: GDP value for the reference year.}

\item{ref_pce}{numeric: PCE value for the reference year.}
}
\value{
numeric
}
\description{
Find the correct extrapolation or interpolation method to use in
\code{\link[=adjust_welfare_mean]{adjust_welfare_mean()}}.
}
\keyword{internal}
