% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_growth_same_direction}
\alias{is_growth_same_direction}
\title{Find whether svy growth and nac growth go in the same direction}
\usage{
is_growth_same_direction(
  survey_mean1,
  survey_mean2,
  svy_value1,
  svy_value2,
  ref_value
)
}
\arguments{
\item{survey_mean1}{numeric: value of survey mean in year 1}

\item{survey_mean2}{numeric: value of survey mean in year 2}

\item{svy_value1}{numeric: value of national accounts in survey year 1}

\item{svy_value2}{numeric: value of national accounts in survey year 2}

\item{ref_value}{numeric: value  of national accounts in reference year}
}
\value{
logical
}
\description{
Find whether svy growth and nac growth go in the same direction
}
