% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_same_direction}
\alias{is_same_direction}
\title{Is growth from survyes the same of NAC}
\usage{
is_same_direction(svy_means, svy_nac, ref_nac)
}
\arguments{
\item{svy_means}{numeric: vector of survey means}

\item{svy_nac}{numeric: vector of National Accounts values}

\item{ref_nac}{numeric: vector of national account value in reference year}
}
\value{
logical vector
}
\description{
checks growth in the survey mean between the two surveys is of the same sign
as (1) the growth in national accounts from the first survey to the reference
year, and (2) from the reference year to the second survey.
}
