% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_to_interpolate}
\alias{is_to_interpolate}
\title{is this lineup year to interpolate?}
\usage{
is_to_interpolate(x)
}
\arguments{
\item{x}{numric: vector with survey years}
}
\value{
logical
}
\description{
is this lineup year to interpolate?
}
