% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dsm_table.R
\name{join_dsm_tables}
\alias{join_dsm_tables}
\title{Append new and old DSM tables}
\usage{
join_dsm_tables(new, old)
}
\arguments{
\item{new}{data.table: A table with newly estimated deflated survey means.}

\item{old}{data.table: The current DSM table. Output of \code{read_dsm()}.}
}
\value{
data.table
}
\description{
Append new and old DSM tables
}
\keyword{internal}
