% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lcu_table.R
\name{lcum}
\alias{lcum}
\title{Calculate Survey mean in LCU for one survey}
\usage{
lcum(inv)
}
\arguments{
\item{inv}{character: survey ID.}
}
\value{

}
\description{
Calculate Survey mean in LCU for one survey
}
