% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_and_clean.R
\name{load_clean}
\alias{load_clean}
\title{Load one data at a time and clean it}
\usage{
load_clean(survey_id, maindir)
}
\arguments{
\item{maindir}{}
}
\value{

}
\description{
Load one data at a time and clean it
}
