% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microdata}
\alias{microdata}
\title{Synthetic microdata}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
\item{country_code}{A fake three letter country code}
\item{survey_year}{The year of the survey}
\item{weight}{The weight of each indvidual's welfare value}
\item{welfare}{The welfare of each indvidual. Can represent either income or consumption.}
\item{survey_coverage}{Geographical location; Urban or Rural}
\item{gender}{Gender; Male or Female}
}
}
\source{
Worldbank internal.
}
\usage{
microdata
}
\description{
A dataset containing synthetic microdata for 2000 individuals in a fake
country.
}
\details{
Note that nothing in this dataset is real. It is only meant for testing and
example purposes.
}
\keyword{datasets}
