% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_paths}
\alias{pip_paths}
\title{PIP paths}
\format{
An object of class \code{list} of length 6.
}
\usage{
pip_paths
}
\description{
A list with default PIP directory and file paths.
}
\details{
\describe{
\item{dlwdir}{Path to Datalibweb directory.}
\item{production_datadir}{Path to the production PIP-data directory.}
\item{production_pipedir}{Path to the production ingestion pipeline director.}
\item{staging_datadir}{Path to the staging PIP-data directory.}
\item{testing_datadir}{Path to the testing PIP-data directory.}
\item{dsmfile}{Path to current DSM file.}
}
}
\keyword{datasets}
