% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_update_cache_inventory.R
\name{pip_update_cache_inventory}
\alias{pip_update_cache_inventory}
\title{Update correspondence inventory file}
\usage{
pip_update_cache_inventory(
  pipeline_inventory = NULL,
  pip_data_dir = gls$PIP_DATA_DIR,
  cache_svy_dir = NULL,
  tool = c("PC", "TB")
)
}
\arguments{
\item{pipeline_inventory}{data.table: Pipeline inventory table.}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cache_svy_dir}{character: Output folder for the cached survey data.}

\item{tool}{character: Either "PC" or "TB" for poverty calculator and Table
Baker, receptively}
}
\value{
TRUE if file is update. FALSE If no data is in directory
}
\description{
Update correspondence inventory file
}
