% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipdm-package.R
\docType{package}
\name{pipdm-package}
\alias{pipdm}
\alias{pipdm-package}
\title{pipdm: PIP Data Managment}
\description{
A package for managing the creation of the PIP database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIP-Technical-Team/pipdm}
  \item \url{https://pip-technical-team.github.io/pipdm}
  \item Report bugs at \url{https://github.com/PIP-Technical-Team/pipdm/issues}
}

}
\author{
\strong{Maintainer}: Tony Fujs \email{tfujs@worldbank.org}

Authors:
\itemize{
  \item Aleksander Eilertsen \email{aeilertsen@worldbank.org}
  \item R.Andrés Castañeda \email{acastanedaa@worldbank.org}
}

}
\keyword{internal}
