% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipdm.R
\docType{package}
\name{pipdm}
\alias{pipdm}
\title{pipdm: A package for computating the notorious bar statistic}
\description{
The pipdm package provides three categories of important functions:
pipdm, bar and baz.
}
\section{pipdm functions}{

The pipdm functions ...
}

