% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ppp_list.R
\name{ppp_filter}
\alias{ppp_filter}
\title{Filter PPP data from using the structure of \code{pipload::pip_load_aux("ppp")}}
\usage{
ppp_filter(ppp, pt, rv, av)
}
\arguments{
\item{ppp}{data frame with ppp value. must have the same strcuture as in
\code{pipload::pip_load_aux("ppp")}}

\item{pt}{numeric: ppp release years}

\item{rv}{character: Release Version in the form "vX" where "X" refers to the
version number}

\item{av}{character: Adaptation Version in the form "vX" where "X" refers to
the version number}
}
\value{
data frmae
}
\description{
Filter PPP data from using the structure of \code{pipload::pip_load_aux("ppp")}
}
