% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_svy_data_to_cache.R
\name{process_svy_data_to_cache}
\alias{process_svy_data_to_cache}
\title{Process survey data to cache file}
\usage{
process_svy_data_to_cache(
  survey_id,
  cache_id,
  pip_data_dir,
  cache_svy_dir,
  compress = 100,
  cols = NULL,
  cpi_table,
  ppp_table,
  pfw_table,
  pop_table
)
}
\arguments{
\item{survey_id}{character: Original Survey ID}

\item{cache_id}{character: cache id vector}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cache_svy_dir}{character: Output directory}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{cols}{character: vector of variables to keep. Default is NULL.}

\item{cpi_table}{data.table: A table with CPI data.}

\item{ppp_table}{data.table: A table with PPP data.}

\item{pfw_table}{data.table: A table with the price framework file.}

\item{pop_table}{data.table: A table with population data.}
}
\value{
data frame with status of process
}
\description{
Process survey data to cache file
}
