% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dsm.R
\name{read_dsm}
\alias{read_dsm}
\title{Read DSM table}
\usage{
read_dsm(path)
}
\arguments{
\item{path}{character: Path to the current DSM file.}
}
\value{
data.table
}
\description{
Load the current DSM table.
}
