% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inventory.R
\name{read_inventory}
\alias{read_inventory}
\title{Read inventory}
\usage{
read_inventory(path)
}
\arguments{
\item{path}{character: Path to the inventory file.}
}
\value{
data.table
}
\description{
Load the raw inventory file.
}
