% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_dir_growth}
\alias{same_dir_growth}
\title{same_dir_growth}
\usage{
same_dir_growth(svy_means, svy_nac, ref_nac)
}
\arguments{
\item{svy_means}{numeric: vector of survey means}

\item{svy_nac}{numeric: vector of National Accounts values}

\item{ref_nac}{numeric: vector of national account value in reference year}
}
\value{
numeric vector
}
\description{
Estimate growth factor for each distribution when growth in means is of the
same direction that growth of NACs to reference year.
}
