% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_aux_data.R
\name{save_aux_data}
\alias{save_aux_data}
\title{Save auxiliary data.}
\usage{
save_aux_data(x, filename, compress = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{filename}{final name of aux data}

\item{compress}{logical. iF TRUE it compress output}
}
\description{
Save auxiliary data.
}
