% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_survey_data.R
\name{save_survey_data}
\alias{save_survey_data}
\title{Save survey data}
\usage{
save_survey_data(
  dt,
  cols = NULL,
  output_dir,
  cache_filename,
  future_plan = c("sequential", "multisession", "callr"),
  compress,
  save = TRUE,
  load = FALSE
)
}
\arguments{
\item{dt}{Data frame from cache to be saved}

\item{cols}{character: Vector with columns to save. If NULL all columns are
saved.}

\item{output_dir}{character: Output folder.}

\item{cache_filename}{character: Vector with new names for microdata.}

\item{future_plan}{character: \code{future} plan to use.}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{save}{logical: if TRUE save data file. Default is TRUE}

\item{load}{logical: if TRUE loads data. Defatul is FALSE and returns file
directory path}
}
\description{
Save survey data to .fst file in specified output directory.
}
