% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_closest_surveys.R
\name{select_closest_surveys}
\alias{select_closest_surveys}
\title{Get all closest surveys to a reference year, for a single welfare type}
\usage{
select_closest_surveys(svy_lkup, ref_year)
}
\arguments{
\item{svy_lkup}{data.frame: Look-up table for survey data.}

\item{ref_year}{integer: The selected reference year.}
}
\value{
data.frame
}
\description{
For any reference year, retrieves information about the closest surveys.
}
\keyword{internal}
