% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_proxy.R
\name{select_proxy}
\alias{select_proxy}
\title{Select proxy values}
\usage{
select_proxy(svy_table, region_code, ref_gdp, ref_pce)
}
\arguments{
\item{svy_table}{data.frame: A table with information on surveys and national
accounts data.}

\item{region_code}{character: World Bank three letter region code.}

\item{ref_gdp}{numeric: GDP value for the reference year.}

\item{ref_pce}{numeric: PCE value for the reference year.}
}
\value{
list
}
\description{
Select GDP or PCE as proxy value based on pre-specified rules.
}
\details{
PCE is the default for all countries, except in Sub-Saharan Africa where GDP
is always used. If any PCE value is missing for the specific reference year,
GDP is used.
}
\references{
Prydz, E.B., D. Jolliffe, C. Lakner, D.G. Mahler, P. Sangraula. 2019.
"\href{http://documents1.worldbank.org/curated/en/664751553100573765/pdf/135460-WP-PUBLIC-Disclosed-3-21-2019.pdf}{National Accounts Data used in Global Poverty Measurement}".
Global Poverty Monitoring Technical Note 8.
World Bank, Washington, DC.
}
\keyword{internal}
