% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_svy_data_to_cache.R
\name{adjust_population}
\alias{adjust_population}
\title{Adjust microdata to WDI population levels when the number of reporting levels
is equal or greater than 2}
\usage{
adjust_population(df, pop_table)
}
\arguments{
\item{df}{dataframe with microdata}

\item{pop_table}{population data from WDI.}
}
\value{
dataframe
}
\description{
Adjust microdata to WDI population levels when the number of reporting levels
is equal or greater than 2
}
