% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_gini}
\alias{compute_gini}
\title{Gini coefficient}
\usage{
compute_gini(y, w)
}
\arguments{
\item{y}{numeric: A vector of income or consumption values}

\item{w}{numeric: A vector of weights}
}
\description{
Compute the Gini coefficient.
}
\details{
Given a vector of income or consumption values (\code{y}) and their respective
weights (\code{w}) \code{compute_gini} computes the Gini coefficient for the
distribution.

Note that the input parameters cannot contain missing values and that
\code{y} must be sorted in increasing order.
}
