% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_lorenz}
\alias{compute_lorenz}
\title{Lorenz curve}
\usage{
compute_lorenz(y, w)
}
\arguments{
\item{y}{numeric: A vector of income or consumption values}

\item{w}{numeric: A vector of weights}
}
\description{
Compute the Lorenz curve.
}
\details{
Given a vector of income or consumption values (\code{y}) and their respective
weights (\code{w}) \code{compute_lorenz} computes the Lorenz curve the
distribution.

Note that the input parameters cannot contain missing values and that
\code{y} must be sorted in increasing order.
}
