% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_mld}
\alias{compute_mld}
\title{Mean log deviation}
\usage{
compute_mld(y, w)
}
\arguments{
\item{y}{numeric: A vector of income or consumption values}

\item{w}{numeric: A vector of weights}
}
\description{
Compute mean log deviation.
}
\details{
Given a vector of income or consumption values (\code{y}) and their respective
weights (\code{w}) \code{compute_mld} computes the mean log deviation for the
distribution.

Note that the input parameters cannot contain missing values and that
\code{y} must be sorted in increasing order.
}
