% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cache_file.R
\name{create_cache_file}
\alias{create_cache_file}
\title{Cache survey data}
\usage{
create_cache_file(
  pipeline_inventory,
  pip_data_dir,
  cache_svy_dir,
  tool = c("PC", "TB"),
  compress = 100,
  verbose = FALSE,
  force = FALSE,
  cpi_table,
  ppp_table,
  pfw_table,
  pop_table
)
}
\arguments{
\item{pipeline_inventory}{data.table: Pipeline inventory table.}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cache_svy_dir}{character: Output folder for the cached survey data.}

\item{tool}{character: Either "PC" or "TB" for poverty calculator and Table
Baker, receptively}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{verbose}{boolean: If TRUE additional messages are printed to the
console.}

\item{force}{logical: if TRUE, all files will be generate again.}

\item{pfw_table}{data.table: A table with the price framework file.}

\item{pop_table}{data.table: A table with population data.}
}
\value{
list: creation status and cache data availability
}
\description{
Save a cleaned version of the survey data in the cache survey data directory.
}
