% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cache_file.R
\name{create_cache_file}
\alias{create_cache_file}
\title{Cache survey data}
\usage{
create_cache_file(
  pipeline_inventory = NULL,
  pip_data_dir = getOption("pip.maindir"),
  tool = c("PC", "TB"),
  cache_svy_dir = NULL,
  compress = 100,
  verbose = TRUE,
  force = FALSE,
  cpi_dt = NULL,
  ppp_dt = NULL
)
}
\arguments{
\item{pipeline_inventory}{data.table: Pipeline inventory table.}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cache_svy_dir}{character: Output folder for the cached survey data.}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{verbose}{boolean: If TRUE additional messages are printed to the
console.}

\item{force}{logical: if TRUE, all files will be generate again.}

\item{cpi_dt}{data frame with CPI data to deflate welfare in TB data}

\item{ppp_dt}{data frame with PPP data to deflate welfare in TB data}
}
\value{
list: creation status and cache data availability
}
\description{
Save a cleaned version of the survey data in the cache survey data directory.
}
