% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clean_data.R
\name{db_clean_data}
\alias{db_clean_data}
\title{Clean survey data}
\usage{
db_clean_data(dt, gc = FALSE)
}
\arguments{
\item{dt}{data.table: A survey dataset.}

\item{gc}{logical: If TRUE garbage collection is forced.}
}
\value{
data.table
}
\description{
Clean a survey datasets by applying methods from \code{wbpip:::md_clean_data()}
and \code{wbpip:::gd_clean_data()}.
}
