% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compute_survey_mean.R
\name{db_compute_survey_mean}
\alias{db_compute_survey_mean}
\title{Calculate survey means}
\usage{
db_compute_survey_mean(dt, gd_mean = NULL, gc = FALSE)
}
\arguments{
\item{dt}{data.table: A survey dataset.}

\item{gd_mean}{numeric: Mean to use for grouped data surveys.}

\item{gc}{logical: If TRUE garbage collection is forced.}
}
\value{
data.table
}
\description{
Calculate survey mean in local currency units (LCU) for a single survey
dataset.
}
