% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dist_table.R
\name{db_create_dist_table}
\alias{db_create_dist_table}
\title{Create distributional statistics table}
\usage{
db_create_dist_table(dl, dsm_table, crr_inv, verbose = FALSE)
}
\arguments{
\item{dl}{list: A list with distributional statistics datasets.}

\item{dsm_table}{data.table: A table with deflated survey means.}

\item{crr_inv}{data frame with correspondence inventory}

\item{verbose}{logical: Whether to display messages. Default is TRUE}
}
\value{
data.table
}
\description{
Create a table with distributional statistics, including a deflated median.
}
