% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dsm_table.R
\name{db_create_dsm_table}
\alias{db_create_dsm_table}
\title{Deflate survey mean in LCU to PPP}
\usage{
db_create_dsm_table(
  inv = NULL,
  cpi = NULL,
  ppp = NULL,
  dt = NULL,
  append = TRUE,
  datadir = getOption("pip.datadir"),
  pipedir = getOption("pip.pipedir")
)
}
\arguments{
\item{inv}{character: Survey IDs to be parsed to \code{db_create_table} in case
argument \code{dt} is NULL. Only works is \code{dt} is NULL.}

\item{cpi}{Data frame with most latest version of CPI}

\item{ppp}{Data frame with most latest version of PPP}

\item{dt}{Updated table with LCU means. If NULL call \code{db_create_lcu_table}
parsing argument \code{inv}}

\item{append}{logical: If TRUE, append to current dsm table file and save.
TRUE is default.}

\item{datadir}{character: PIP-Data directory}

\item{pipedir}{character: Pipeline ingestion directory}
}
\value{

}
\description{
Deflate survey mean in LCU to PPP
}
