% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lcu_table.R
\name{db_create_lcu_table}
\alias{db_create_lcu_table}
\title{Create survey mean table}
\usage{
db_create_lcu_table(dl)
}
\arguments{
\item{dl}{list: A list with survey datasets. See details.}
}
\value{
data.table
}
\description{
Create a table with welfare means in Local Currency Units (LCU) for each survey.
}
\details{
The supplied survey datasets  \emph{must} include the following columns;
\code{survey_id}, \code{welfare}, \code{weight}, \code{cpi_data_level} and \code{ppp_data_level}.
}
