% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lcu_table.R
\name{db_create_lcu_table}
\alias{db_create_lcu_table}
\title{Create survey mean table (LCU)}
\usage{
db_create_lcu_table(dlc, pop, maindir)
}
\arguments{
\item{dlc}{dataframe with welfare data loaded and clened from
\code{db_load_and_clean}.}

\item{pop}{data.table: A table with population data.}
}
\value{
data.table
}
\description{
Create a table with welfare means in Local Currency Units (LCU) for each
survey.
}
