% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_create_lkup_table}
\alias{db_create_lkup_table}
\title{Create look-up table}
\usage{
db_create_lkup_table(df, nac_table, ref_years)
}
\arguments{
\item{df}{data.table: Output of \code{\link[=db_adjust_nac_values]{db_adjust_nac_values()}}.}

\item{nac_table}{data.table: Output of \code{\link[=db_subset_nac_table]{db_subset_nac_table()}}.}

\item{ref_years}{numeric: Vector with reference years.}
}
\value{
\code{data.table}
}
\description{
Create a look-up table with information on all reference years and surveys
for further data manipulation.
}
\keyword{internal}
