% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_lkup_table.R
\name{db_create_lkup_table}
\alias{db_create_lkup_table}
\title{Create reference year look-up table}
\usage{
db_create_lkup_table(dt, nac_table, pop_table, ref_years, region_code)
}
\arguments{
\item{dt}{data.table: A table with survey data. See details.}

\item{nac_table}{data.table: A table with GDP and PCE data. Output of
\code{\link[=db_create_nac_table]{db_create_nac_table()}}.}

\item{pop_table}{data.table: A table with population data.}

\item{ref_years}{numeric: A vector with reference years.}

\item{region_code}{character: A value with the region code column to use for
calculating predicted means. Either 'pcn_region_code' or 'wb_region_code'.}
}
\value{
data.table
}
\description{
Create a look-up table with information on all reference years and surveys
for further data manipulation.
}
\details{
\code{dt} must included information on survey metadata, survey national account
values and deflated survey means.
}
\keyword{internal}
