% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_nac_table.R
\name{db_create_nac_table}
\alias{db_create_nac_table}
\title{Merge GDP and PCE data}
\usage{
db_create_nac_table(gdp_table, pce_table)
}
\arguments{
\item{gdp_table}{data.frame: A table with GDP data.}

\item{pce_table}{data.frame: A table with PCE data.}
}
\value{
\code{data.table}
}
\description{
Create a table with national accounts (NAC) data by merging separate tables
for GDP and PCE.
}
\details{
Both input tables should contain ISO-3 country codes, calender years and
information indicating the data level and domain (e.g national, urban/rural
or sub-national) of the variable in question. Note that data levels must be
comparable across datasets.
}
\keyword{internal}
