% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_pipeline_inventory.R
\name{db_create_pipeline_inventory}
\alias{db_create_pipeline_inventory}
\title{Create pipeline inventory. This is different from the pip inventory. PIP
inventory is the surveys available in the PIP repository. Pipeline inventory
takes is the PIP inventory, filtered by the Price FrameWork dataset}
\usage{
db_create_pipeline_inventory(
  root_dir = Sys.getenv("PIP_ROOT_DIR"),
  maindir = gsl$PIP_DATA_DIR,
  inv_file = paste0(maindir, "_inventory/inventory.fst")
)
}
\arguments{
\item{root_dir}{character: root directory of the PIP data}

\item{maindir}{character: Main directory}

\item{inv_file}{character: file path to be loaded.}
}
\value{

}
\description{
Create pipeline inventory. This is different from the pip inventory. PIP
inventory is the surveys available in the PIP repository. Pipeline inventory
takes is the PIP inventory, filtered by the Price FrameWork dataset
}
\examples{
db_create_pipeline_inventory()
}
