% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_create_ref_year_table}
\alias{db_create_ref_year_table}
\title{Create reference year table}
\usage{
db_create_ref_year_table(
  dsm_table,
  gdp_table = pipload::pip_load_aux("gdp"),
  pce_table = pipload::pip_load_aux("pce"),
  pop_table = pipload::pip_load_aux("pop"),
  pfw_table = pipload::pip_load_aux("pfw"),
  ref_years = gls$PIP_REF_YEARS,
  pip_years = gls$PIP_YEARS,
  region_code = c("pcn_region_code", "wb_region_code")
)
}
\arguments{
\item{dsm_table}{data.table: A table with deflated survey means.}

\item{gdp_table}{data.table: A table with GDP data.}

\item{pce_table}{data.table: A table with PCE data.}

\item{pop_table}{data.table: A table with population data.}

\item{pfw_table}{data.table: A table with the price framework file.}

\item{ref_years}{numeric: A vector with reference years.}

\item{pip_years}{numeric: A vector with calender years used in PIP.}

\item{region_code}{character: A value with the region code column to use.}
}
\value{
data.table
}
\description{
Create a table with adjusted welfare means for each reference year.
}
\details{
\code{db_create_ref_year_table()} creates a table with interpolated or
extrapolated welfare means for all specified reference years based on the
provided survey, GDP and PCE data.
}
