% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_create_ref_year_table}
\alias{db_create_ref_year_table}
\title{Create reference year table}
\usage{
db_create_ref_year_table(
  gdp_table,
  pce_table,
  pop_table,
  pfw_table,
  dsm_table,
  ref_years,
  pip_years
)
}
\arguments{
\item{gdp_table}{data.table: A table with GDP data.}

\item{pce_table}{data.table: A table with PCE data.}

\item{pop_table}{data.table: A table with population data.}

\item{pfw_table}{data.table: A table with the price framework file.}

\item{dsm_table}{data.table: A table with deflated survey means.}

\item{ref_years}{numeric: A vector with reference years.}

\item{pip_years}{numeric: A vector with calender years used in PIP.}
}
\value{
data.table
}
\description{
Create a table with adjusted welfare means for each reference year.
}
\details{
\code{db_create_ref_year_table()} creates a table for with interpolated or
extrapolated welfare means for all specified reference years based on the
provided survey, GDP and PCE data.
}
