% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_filter_inventory.R
\name{db_filter_inventory}
\alias{db_filter_inventory}
\title{Filter inventory}
\usage{
db_filter_inventory(
  raw_inventory = getOption("pip.inventory"),
  dsm_file = getOption("pip.dsmfile"),
  datadir = getOption("pip.datadir"),
  pipedir = getOption("pip.pipedir")
)
}
\arguments{
\item{raw_inventory}{character: Path to the raw inventory. Defaults to
\code{getOption("pip.inventory")}.}

\item{dsm_file}{character: Path to the DSM file. Defaults to
\code{getOption("pip.dsmfile")}}

\item{datadir}{character: Path to the PIP-Data directory. Defaults to
\code{getOption("pip.datadir")}.}

\item{pipedir}{character: Path to the ingestion pipeline directory. Defaults
to \code{getOption("pip.pipedir")}.}
}
\value{
data.table
}
\description{
Filter the PIP-Data inventory for those surveys that are not in the existing
DSM table.
}
