% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_filter_inventory.R
\name{db_filter_inventory}
\alias{db_filter_inventory}
\title{Filter inventory}
\usage{
db_filter_inventory(dt, pfw_table)
}
\arguments{
\item{dt}{data.table: A table with the inventory.}

\item{pfw_table}{data.table: A table with the price framework file.}
}
\value{
data.table
}
\description{
Filter the PIP-Data inventory to select surveys in the Price Framework.
}
