% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_filter_inventory.R
\name{db_filter_inventory}
\alias{db_filter_inventory}
\title{filter repository of those surveys that are already in used in dsm_table}
\usage{
db_filter_inventory(
  raw_inventory = getOption("pip.inventory"),
  dsm_file = getOption("pip.dsmfile"),
  datadir = getOption("pip.datadir"),
  pipedir = getOption("pip.pipedir")
)
}
\arguments{
\item{raw_inventory}{dataframe from \code{getOption("pip.inventory")}}

\item{dsm_file}{character: dsm table file. Dafault \code{getOption("pip.dsmfile")}}

\item{datadir}{character: PIP-Data directory}

\item{pipedir}{character: Pipeline ingestion directory}
}
\value{

}
\description{
filter repository of those surveys that are already in used in dsm_table
}
