% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_filter_inventory.R
\name{db_filter_inventory}
\alias{db_filter_inventory}
\title{Filter inventory}
\usage{
db_filter_inventory(raw_inventory, dsm_table)
}
\arguments{
\item{raw_inventory}{data.table: A table with the raw inventory. Output of
\code{read_inventory()}.}

\item{dsm_table}{data.table: A table with deflated survey means. Output of
\code{read_dsm()}.}
}
\value{
data.table
}
\description{
Filter the PIP-Data inventory for those surveys that are not in the existing
DSM table.
}
