% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_finalize_ref_year_table.R
\name{db_finalize_ref_year_table}
\alias{db_finalize_ref_year_table}
\title{Finalize reference year table}
\usage{
db_finalize_ref_year_table(dt, svy_anchor)
}
\arguments{
\item{dt}{data.table: Output of \code{\link[=db_compute_predicted_means]{db_compute_predicted_means()}}.}

\item{svy_anchor}{data.table: A table with survey metadata information.}
}
\value{
data.table
}
\description{
Finalize the reference year table by subsetting the correct rows and
columns.
}
\keyword{internal}
