% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_merge_anchor_nac}
\alias{db_merge_anchor_nac}
\title{Merge survey anchor with national accounts data}
\usage{
db_merge_anchor_nac(svy_anchor, nac_table)
}
\arguments{
\item{svy_anchor}{data.frame: A table with survey metadata information.}

\item{nac_table}{data.table: Output of \code{\link[=db_subset_nac_table]{db_subset_nac_table()}}.}
}
\value{
\code{data.table}
}
\description{
Create a combined table of survey metadata information and national accounts
data.
}
\keyword{internal}
