% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_ref_year_table.R
\name{db_select_lineup_surveys}
\alias{db_select_lineup_surveys}
\title{Select correct line-up surveys}
\usage{
db_select_lineup_surveys(df)
}
\arguments{
\item{df}{data.table: Output of \code{\link[=db_adjust_welfare_mean]{db_adjust_welfare_mean()}}.}
}
\value{
\code{data.table}
}
\description{
Select a single line-up survey when there are multiple line-up surveys
available.
}
\seealso{
\code{\link[=select_lineup_survey]{select_lineup_survey()}}
}
\keyword{internal}
