% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dsm_table.R
\name{join_dsm_tables}
\alias{join_dsm_tables}
\title{Append new and old DSM tables}
\usage{
join_dsm_tables(new, old = NULL)
}
\arguments{
\item{new}{data.table: A table from \code{db_create_dsm_table()}.}

\item{old}{data.table: A table from \code{getOption("pip.dsmfile")}.}
}
\value{
data.table
}
\description{
Append new and old DSM tables
}
\keyword{internal}
