% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pip_paths}
\alias{pip_paths}
\title{PIP paths}
\format{
An object of class \code{list} of length 5.
}
\usage{
pip_paths
}
\description{
A list with default PIP directory and file paths.
}
\details{
\describe{
\item{dlwdir}{Path to Datalibweb directory}
\item{datadir}{Path to PIP-data directory}
\item{pipedir}{Path to PIP ingestion pipeline directory.}
\item{dsmfile}{Path to current DSM file.}
\item{inventory}{Path to raw inventory file.}
}
}
\keyword{datasets}
