% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_svy_data_to_cache.R
\name{process_svy_data_to_cache}
\alias{process_svy_data_to_cache}
\title{Process survey data to cache file}
\usage{
process_svy_data_to_cache(
  survey_id,
  cache_id,
  pip_data_dir = gls$PIP_DATA_DIR,
  cache_svy_dir = gls$CACHE_SVY_DIR_PC,
  compress = 100,
  cols = NULL,
  cpi_dt = pipload::pip_load_aux("cpi"),
  ppp_dt = pipload::pip_load_aux("ppp"),
  pfw_dt = pipload::pip_load_aux("pfw")
)
}
\arguments{
\item{survey_id}{character: Original Survey ID}

\item{cache_id}{character: cache id vector}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cache_svy_dir}{character: Output directory}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}

\item{cols}{character: vector of variables to keep. Default is NULL.}

\item{cpi_dt}{data frame with CPI data to deflate welfare}

\item{ppp_dt}{data frame with PPP data to deflate welfare}

\item{pfw_dt}{data frame with Price FrameWork data}
}
\value{
data frame with status of process
}
\description{
Process survey data to cache file
}
