% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_svy_data_to_cache.R
\name{process_svy_data_to_cache}
\alias{process_svy_data_to_cache}
\title{Process survey data to cache file}
\usage{
process_svy_data_to_cache(
  survey_id,
  chh_filename,
  pip_data_dir,
  cols = NULL,
  cache_svy_dir,
  compress
)
}
\arguments{
\item{survey_id}{character: Original Survey ID}

\item{chh_filename}{character: Cache filename}

\item{pip_data_dir}{character: Input folder for the raw survey data.}

\item{cols}{character: vector of variables to keep. Default is NULL.}

\item{cache_svy_dir}{character: Output directory}

\item{compress}{numeric: Compression level used in \code{fst::write_fst()}.}
}
\value{

}
\description{
Process survey data to cache file
}
