% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_dsm_table.R
\name{save_dsm}
\alias{save_dsm}
\title{Save DSM table}
\usage{
save_dsm(new_dsm)
}
\arguments{
\item{new_dsm}{data.table: A data frame containing both the old and newly
estimated DSM tables. Output of \code{join_dsm_tables()}.}
}
\value{
logical
}
\description{
Save the deflated survey mean table to multiple locations.
}
\keyword{internal}
