% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dsm.R
\name{save_dsm}
\alias{save_dsm}
\title{Save DSM table}
\usage{
save_dsm(new_dsm, path)
}
\arguments{
\item{new_dsm}{data.table: A data frame containing both the old and newly
estimated DSM tables. Output of \code{join_dsm_tables()}.}

\item{path}{character: Path to the DSM file.}
}
\value{
logical
}
\description{
Save the deflated survey mean table to multiple locations.
}
