% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniq_vars_to_attr}
\alias{uniq_vars_to_attr}
\title{convert variagles with unique values along the data set to attrbitus and then
remove those unique variables}
\usage{
uniq_vars_to_attr(x)
}
\arguments{
\item{x}{data frame.}
}
\description{
convert variagles with unique values along the data set to attrbitus and then
remove those unique variables
}
